.ipynb_checkpoints/requirements-checkpoint.txt
numpy==1.22.2
Pillow==9.5.0
streamlit==1.26.0
tensorflow==2.13.0rc1
tensorflow_macos==2.13.0rc1